[server] create evaluation package
